[EX 2] Add the base code given on Moodle
The make all failed for some reasons:
- No bin/ directory was present, it is now created
- The iterative src code doesn't exist yet. The target to compile it has
  been commented in the all target.
- Some weird behaviour with lapack.h.
  The call to the dgbtrs function was possible because there was to few
arguments.
  It turns out that the lapack.h installed on my computer defined a
FORTRAN constant that modify the signature of some function adding
a last argument (size) to the dgbtrs routine for example.
  The define has been commented (temporary solution) so the functions
defined in the header and the ones present in the doc have the same signature.
- The ambre.mk was renamed archlinux.mk and the -lblas option was
  changed to -lcblas in order to use the right library.

The make all command now runs perfectly and the binaries generated too.
The testenv seems to print right results.