[Ex 4] Added the set functions
The function called to set the matrix in a column major and general band
form, set the X vector used to compute the analytic solution, set the
rirht hand member of the equation and compute and set the EX_XOL vector
with the analytic solution.

The writting of the different arrays has been uncommented to verify the
set functions.