[Ex 4] Fix set matrix, add resolution and error
The set_GB_operator_colMajor_poisson1D didn't set the kv lines correctly
leading to false results when lapack was invoked. The kv lines was after
the data instead of before. This was due to a misreading of the
documentation.

The main function now resolves the heat equation with a direct method
and the analytic one. The forward error is compute and printed.

The dgbsv subroutine has been added and the forward error with this
fonction is also computed and printed.