[EX 5-6] LAPACK functions and tridiagonal impl
Firstly, the Makefile and archlinux.mk files have been modified in order
to have a better optimisation for the specific architecture I am on.

Secondly, the exercise 5 is implemented. So, the DGBTRF, DGBTRS and
DGBSV subroutines are now used in the code to find a solution of our
problem.

Then, the tridiagonal specific triangularization algorithm have been
implemented in the form of a DGBTRFTRIDIAG functions that follows some
of the convention of LAPACK subroutines.

Finally, a precise measure of performance has been implemented using
one of the hardware clocks of the computer. The measure is performed
uniquely on the kernel computations. The size of the problem/matrix has
been increased to have a better relative error on the measurement of the
performance.