Add time measurements and plotting
-Added warning flags to help clean the code.
-Added the plot generated to the clean target.

All the methods are now measured in function of the size of the entry.
The data collected are written down in dat files.
Those data can be plot using the gnuplot scripts located in the plot
directory.

A strange behaviour (that may not be one) happens with the number of
iterations of the different methods. This one increases with the size of
the entry until a certain value that is not the maximum number of
iterations. This number then stays steady.