Final restructuration
Add error handling to all functions. The allocation are now on checked
before use. The libpoisson functions return errors accordingly with the
type of error (bad argument mostly).

Change the computation of eigvalues to be the same as the one seen in
the course rather than the one seen in TP. This is the method exposed in
the report and used to prove the optimum alpha.

Removes most of the warnings returned by the compiler. It was mainly
unused variables.

Format the code: rework comments, change values to have a quick
execution, change display for the user...

The plots are modified to create french graphics. They are also fixed,
the path to find the data was not changed when they were moved to their
folder.